+++ tom_ast_model/lib/src/ast/ast_core.dart
@@ -445,6 +445,7 @@ void _registerAllFactories() {
     'RecordTypeAnnotation',
     SRecordTypeAnnotation.fromJson,
   );
+  SAstNodeFactory.register('RecordTypeField', SRecordTypeField.fromJson);
 
   // Function-related
   SAstNodeFactory.register(